Split out old ProtoCompat module into several modules. Wrote name validation parsers. Updated ProtoCompat types & IsMessage class for modified .proto file. Wrote isomorphism between names (for convenience/TH). Ported TemplateHaskell utility from other PR.
